Split db tests per dao
