+++ app/src/androidTest/kotlin/com/pitchedapps/frost/db/CookieDbTest.kt
@@ -1,37 +1,11 @@
 package com.pitchedapps.frost.db
 
-import android.content.Context
-import androidx.room.Room
-import androidx.test.core.app.ApplicationProvider
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import kotlinx.coroutines.runBlocking
-import org.junit.runner.RunWith
-import kotlin.test.AfterTest
-import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertNull
 
-@RunWith(AndroidJUnit4::class)
-class FrostDatabaseTest {
-
-    private lateinit var db: FrostDatabase
-
-    @BeforeTest
-    fun before() {
-        val context = ApplicationProvider.getApplicationContext<Context>()
-        val privateDb = Room.inMemoryDatabaseBuilder(
-            context, FrostPrivateDatabase::class.java
-        ).build()
-        val publicDb = Room.inMemoryDatabaseBuilder(
-            context, FrostPublicDatabase::class.java
-        ).build()
-        db = FrostDatabase(privateDb, publicDb)
-    }
-
-    @AfterTest
-    fun after() {
-        db.close()
-    }
+class CookieDbTest : BaseDbTest() {
 
     @Test
     fun basicCookie() {
@@ -61,7 +35,7 @@ class FrostDatabaseTest {
     }
 
     @Test
-    fun insertCookie() {
+    fun insertReplaceCookie() {
         val cookie = CookieEntity(id = 1234L, name = "testName", cookie = "testCookie")
         runBlocking {
             db.cookieDao().insertCookie(cookie)
@@ -78,6 +52,16 @@ class FrostDatabaseTest {
                 db.cookieDao().selectAll().toSet(),
                 "New cookie insertion failed"
             )
+        }
+    }
+
+    @Test
+    fun selectCookie() {
+        val cookie = CookieEntity(id = 1234L, name = "testName", cookie = "testCookie")
+        runBlocking {
+            db.cookieDao().insertCookie(cookie)
+            assertEquals(cookie, db.cookieDao().selectById(cookie.id), "Cookie selection failed")
+            assertNull(db.cookieDao().selectById(cookie.id + 1), "Inexistent cookie selection failed")
         }
     }
 }
